Fixed the wrong playercard appearing
Previously, only the soldiers playercard would display in the selection
screen. Now the correct card is displayed. It should be noted that this
fix is likely suboptimal, and someone should probably look into
optimizing this (i believe my solution causes it to load the image each
refresh, which is probably not ideal)